Reservation toegevoegd

+++ Plathe/Models/Ticket.cs
@@ -9,7 +9,7 @@ namespace Plathe.Models
         public int ShowID { get; set; }
         public string UniqueCode { get; set; }
         public string SeatNumber { get; set; }
-        public int Price { get; set; }
+        public Decimal Price { get; set; }
         public string Options { get; set; }
         public Boolean PopcornTime { get; set; }
         public DateTime PrintedOn { get; set; }
